refactor: code 관리 비지니스 로직 customHook으로 이전
